+++ R/utils.R
@@ -109,3 +109,7 @@ last <- function(x) {
 str_compact <- function(x) {
   x[x != ""]
 }
+
+vec_index_is_empty <- function(x) {
+  !length(x) || all(x == 0L)
+}